add report_html, responsible user in table, refactor structure in raw_materials_report module
