+++ raw_materials_report/models/__init__.py
@@ -1,3 +1,2 @@
 from . import stock_report
 from . import aggregated_foreman_report
-from . import raw_materials_report_wizard
